+++ CHANGELOG.md
@@ -4,9 +4,11 @@ seleniumtid changelog
 v0.9.3
 ------
 
-*Release date: 2015-06-15*
+*Release date: 2015-07-24*
 
   * Fix set_value and app_strings errors in mobile web tests
+  * Allow to set custom driver capabilities from properties file
+    New config section [Capabilities] with driver capabilities
 
 v0.9.2
 ------